requirements.txt
Flask==3.0.3
Pillow==10.4.0
numpy==1.26.4
cryptography==42.0.8
opencv-python-headless==4.10.0.84
gunicorn==22.0.0